完成QuartzFun

+++ QuartzFun/QuartzFun/UIColor+XYRandomColor.m
@@ -15,9 +15,9 @@
         sended = YES;
         srandom(time(NULL));
     }
-    CGFloat red = random() / RAND_MAX;
-    CGFloat green = random() / RAND_MAX;
-    CGFloat blue = random() / RAND_MAX;
+    CGFloat red = (CGFloat)random() / (CGFloat)RAND_MAX;
+    CGFloat green = (CGFloat)random() / (CGFloat)RAND_MAX;
+    CGFloat blue = (CGFloat)random() / (CGFloat)RAND_MAX;
     
     return [UIColor colorWithRed:red green:green blue:blue alpha:1.0f];
 }
